requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.1.1
requests==2.31.0
mysql-connector-python==9.4.0
